create interface in service lib instead
